+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 - ✨(backend) support creating subdoc from file #1987
 - ✨(buildpack) add PaaS deployment support, tested with Scalingo #2293
+- 🔧(backend) allow configuring settings OIDC_OP_USER_ENDPOINT_FORMAT
 
 ### Fixed